requirements.txt
colorama==0.4.6
filelock==3.19.1
fsspec==2025.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
pillow==11.3.0
PyYAML==6.0.2
sympy==1.14.0
torch==2.8.0+cu129
torchvision==0.23.0+cu129
tqdm==4.67.1
typing_extensions==4.14.1
